+++ Block/Adminhtml/Entity/Edit/Form/Renderer/Rating.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Copyright © Alekseon sp. z o.o.
+ * http://www.alekseon.com/
+ */
+namespace Alekseon\AlekseonEav\Block\Adminhtml\Entity\Edit\Form\Renderer;
+
+/**
+ *
+ */
+class Rating extends \Magento\Backend\Block\Template
+{
+    protected $_template = 'Alekseon_AlekseonEav::entity/edit/form/rating.phtml';
+}